extract ReadSelfMySQLThrottleMetric outside of Throttler
